Added a more sensible stopping mechanism.
All accessories now have a stop method that is called before the
AccessoryDriver shuts down. Also, the accessory has a sentinel
threading.Event that can be used to signal the stopping. Thus, to sum up,
when stopping the AccessoryDriver does:
>>> sentinel.set() # Notifies Accessory thread
>>> accessory.stop() # Allows Accessory to clean up resources

The sentinel can be used as:
def run(self):
   while not self.run_sentinel.wait(3):
     do_stuff_every_3_sec_until_set()

Another thing: The AccessoryDriver can be registered as a signal handler in
order to stop itself gracefully. Hit Ctrl+C for example and as if stop is
called.

Nits:
- Improved documentation.
- Cleaned up some code.
- Minor bug fix around server close.
- main.py much simpler and sensible