Fix race condition in publishing value updates.
This change separates the publishing of value changes to clients in two parts - (1)
formatting the change and putting it in a FIFO queue and (2) reading a
change from the FIFO queue and actually sending it to the client. The
latter is done by a dedicated dispatch thread and the abstraction allows
a pool of threads to be adopted in the future (should there be not
enough throughput)

In this way we achieve (1) the thread executing the Accessory's run method is not
blocked while the sending is in progress, (2) when multiple Accessories
(as in the case with a Bridge) publish, there is no race condition, as
adding/reading from the queue is thread-safe.

This change also fixes a bug preventing the access to characteristics of
the Bridge itself.

Finally, this change introduces a socket timeout for the sending of value
changes.