Use HomeKit-defined Characteristic ValidValues.
This change adds a check when assigning a new value to a Characteristic
(using the set_value method) whether the value is in the set of valid
values which are defined for this Characteristic type. For example,
the RotationDirection characteristic defines only 0 and 1 as valid
values.

This change also adds a script, which can be used to parse the
HomeKit-defined Characteristic and Service types (this produces the
files in the resources folder).