Make Http more flexible.
Http is now a Bridge and allows many remote accessories to send updates
with HTTP POST requests. The remote accessories can have optional
characteristics, name, etc. and its state is also persisted.

This change also has minor code changes in the default arguments for
Accessory and Bridge and bumps the version to 1.0.5.