API to bumb conf version #13; Persist on stop #11
This change adds a method to Accessory and AccessoryDriver for notifying
about a configuration change, such as the addition of a service or the
change of a characteristic such as Name. This requires the mDNS info
to be updates, so that iOS clients know they need to fetch new data.

The other change adds a call to AccessoryDriver.persist in
AccessoryDriver.stop - persist the state of the accessory on shutdown.