Persist accessory state in a non-pickle format #12
This change introduces the AccessoryEncoder interface, which an
AccessoryDriver can use to store and load the state of an Accessory. To
recall, the state contains the accessory public and private keys, the
public key and UUID of all paired clients, etc. and it is needed to
support the correct operation of the Accessory after a restart.

The default encoder - AccessoryEncoder, stores the minimum required
state in a json format. The encoder can be changed by passing a
different encoder instance to the AccessoryDriver.

Impact on existing users: If you have a "*.pickle" accessory state and
do nothing, what will probably happen is that your code will correctly
unpickle right after the upgrade to version 1.1.2. On the next restart,
however, the code will crash, as the AccessryDriver will have overriden
the pickle with a non-pickle state representation. At this point, just
don't unpickle as shown in main.py. Alternatively, run
pickle_to_state.py to convert your state.