Allow for both sync and async run method.
This change moves to a cleaner use of asyncio. Introduced
Accessory.start which creates a task for Accessory.run. The idea is to
maintain compatibility with clients that cannot support coroutines and
to allow Accessories to be run both as a thread and as a coroutine.

This is still work in progress.