[WIP] Move to coroutines for Accessory.run (#74)
* Move to coroutines for Accessory.run

* Add support for non-async run methods.

* Allow for both sync and async run method.

This change moves to a cleaner use of asyncio. Introduced
Accessory.start which creates a task for Accessory.run. The idea is to
maintain compatibility with clients that cannot support coroutines and
to allow Accessories to be run both as a thread and as a coroutine.

This is still work in progress.

* Introduce AsyncAccessory, tidying asyncio integ.

* run_at_interval for sync accessories; readme

* Doc changes, removed unused imports