NeoPixelsLightStrip (#144)
* NeoPixelsLightStrip

Added ability to interface with NeoPixels RGB LEDs. This will allow changing of Color, Brightness and State(On/Off) of the NeoPixels attached to the RPi.
This accessory uses all available HomeKit Lightbulb API services (Hue, Saturation, Brightness).
This commit requires the installation of rpi_ws281x library. The procedure is outlined in # https://learn.adafruit.com/neopixels-on-raspberry-pi/software.

* Added check for RGB Tuple length to prevent crash from accessing outside bounds.