Correctly handle stream process stdout/err.
The default Camera.start method now starts a process and
redirects stdout to subprocess.DEVNULL and stderr to a
subprocess.PIP. Also, on Camera.stop, the process is
terminated (was killed) with a timeout of 2 seconds. If
this expires, the process is killed. The stderr of the
process is logged before the terminate.

This change also documents how a camera accessory can be set up
in the README.