Temp disable 'update_advertisment' after pairing (#168)
In some cases, zeroconf does not correctly unregister the service and a subsequent registering causes a NonUniqueName exception. For these cases, clients can set `driver.safe_mode`, which will not cause the driver to skip the call to update_advertisement. The client will need to restart HAP-python in order to get the new advertisement.