Fix random disconnect after upgrade to encrypted (#253)
After server restarts accessories would sometimes go
unavailable because after reconnecting and switching
to encrypted transport part of the unencrypted response
would not have been sent down the wire.  After the
upgrade to encrypted happens that response gets lost
which means the server starts sending encrypted data
when the hap client/controller is still expecting
to finish reading the unencrypted data.

The net result is that the controller/client has to
try again to get another connection to the server
where it does not get unluckly and hit this race
condition.

Before the change "Cleaning connection" was
very common in the log.  Its now rare.

Now that Apple has published more information,
we can improve conformance to HAP spec.

Content-Length is no longer sent on empty content
as it could lead to a loop on the client/controlller side

The Connection, Date, and Server headers are not expected to be
sent according to the spec, and they are now suppressed.