Only send latest state if there are multiple for the same AID/IID in events
iOS 15 had a breaking change where events are no longer processed
in order. We want to make sure when we send an event message we
only send the latest state and overwrite all the previous states
for the same AID/IID that are in the queue when the state changes
before the event is sent.