Allow invalid client values when enabled
As of iOS 15.1, Siri requests TargetHeatingCoolingState
as Auto reguardless if its a valid value or not.

Consumers of this api may wish to set allow_invalid_client_values
to True and handle converting the Auto state to Cool or Heat
depending on the device.

Reference: https://github.com/home-assistant/core/issues/60203